+++ db/seeds.rb
@@ -14,239 +14,14 @@ require "database_cleaner"
 # rubocop:disable Metrics/MethodLength
 # rubocop:disable Metrics/AbcSize
 class SeedDB
-  def initialize
-    @legacy_appeals = []
-    @tasks = []
-    @users = []
-    @ama_appeals = []
-  end
-
-  def create_legacy_appeals(number)
-    legacy_appeals = Array.new(number) do |i|
-      Generators::LegacyAppeal.create(
-        vacols_id: "vacols_id#{i}",
-        vbms_id: "vbms_id#{i}",
-        vacols_record: {
-          status: "Remand"
-        }
-      )
-    end
-
-    @legacy_appeals.push(*legacy_appeals)
-    @legacy_appeals.push(LegacyAppeal.create(vacols_id: "reader_id1", vbms_id: "reader_id1"))
-  end
 
   ### Hearings Setup ###
-  def create_ama_hearing(day)
-    veteran = FactoryBot.create(
-      :veteran,
-      first_name: Faker::Name.first_name,
-      last_name: Faker::Name.last_name
-    )
-
-    appeal = FactoryBot.create(
-      :appeal,
-      veteran_file_number: veteran.file_number,
-      claimants: [FactoryBot.create(:claimant, participant_id: "CLAIMANT_WITH_PVA_AS_VSO")],
-      docket_type: Constants.AMA_DOCKETS.hearing
-    )
-
-    root_task = FactoryBot.create(:root_task, appeal: appeal)
-    distribution_task = FactoryBot.create(
-      :distribution_task,
-      appeal: appeal,
-      parent: root_task
-    )
-    parent_hearing_task = FactoryBot.create(
-      :hearing_task,
-      parent: distribution_task,
-      appeal: appeal
-    )
-
-    FactoryBot.create(
-      :schedule_hearing_task,
-      :completed,
-      parent: parent_hearing_task,
-      appeal: appeal
-    )
-    FactoryBot.create(
-      :assign_hearing_disposition_task,
-      :in_progress,
-      parent: parent_hearing_task,
-      appeal: appeal
-    )
-
-    TrackVeteranTask.sync_tracking_tasks(appeal)
-
-    hearing = FactoryBot.create(
-      :hearing,
-      hearing_day: day,
-      appeal: appeal,
-      bva_poc: User.find_by_css_id("BVAAABSHIRE").full_name,
-      scheduled_time: "9:00AM"
-    )
-
-    FactoryBot.create(
-      :hearing_task_association,
-      hearing: hearing,
-      hearing_task: parent_hearing_task
-    )
-  end
-
-  def create_case_hearing(day, ro_key)
-    case ro_key
-    when "RO17"
-      folder_nr = "3620725"
-    when "RO45"
-      folder_nr = "3411278"
-    when "C"
-      folder_nr = "3542942"
-    end
-
-    FactoryBot.create(
-      :case_hearing,
-      folder_nr: folder_nr,
-      vdkey: day.id,
-      board_member: User.find_by_css_id("BVAAABSHIRE").vacols_attorney_id.to_i
-    )
-  end
-
-  def create_hearing_days
-    user = User.find_by(css_id: "BVATWARNER")
-
-    # Set the current user var here, which is used to populate the
-    # created by field.
-    RequestStore[:current_user] = user
-
-    %w[C RO17 RO19 RO31 RO43 RO45].each do |ro_key|
-      (1..5).each do |index|
-        day = HearingDay.create!(
-          regional_office: (ro_key == "C") ? nil : ro_key,
-          room: "1",
-          judge: User.find_by_css_id("BVAAABSHIRE"),
-          request_type: (ro_key == "C") ? "C" : "V",
-          scheduled_for: Time.zone.today + (index * 11).days,
-          created_by: user,
-          updated_by: user
-        )
-
-        case index
-        when 1
-          create_ama_hearing(day)
-        when 2
-          create_case_hearing(day, ro_key)
-        when 3
-          create_case_hearing(day, ro_key)
-          create_ama_hearing(day)
-        end
-      end
-    end
-
-    # The current user var should be set to nil at the start of this
-    # function. Restore it before executing the next seed function.
-    RequestStore[:current_user] = nil
-  end
 
   ### End Hearings Setup ###
-
-  def create_dispatch_tasks(number)
-    num_appeals = @legacy_appeals.length
-    tasks = Array.new(number) do |i|
-      establish_claim = EstablishClaim.create(
-        appeal: @legacy_appeals[i % num_appeals],
-        aasm_state: :unassigned,
-        prepared_at: rand(3).days.ago
-      )
-      establish_claim
-    end
-
-    # creating user quotas for the existing team quotas
-    team_quota = EstablishClaim.todays_quota
-    UserQuota.create(team_quota: team_quota, user: @users[3])
-    UserQuota.create(team_quota: team_quota, user: @users[4])
-    UserQuota.create(team_quota: team_quota, user: @users[5])
-
-    # Give each user a task in a different state
-    tasks[0].assign!(@users[0])
-
-    tasks[1].assign!(@users[1])
-    tasks[1].start!
-
-    tasks[2].assign!(@users[2])
-    tasks[2].start!
-    tasks[2].review!
-    tasks[2].complete!(status: :routed_to_arc)
-
-    # assigning and moving the task to complete for
-    # user at index 3
-    5.times do |_index|
-      task = EstablishClaim.assign_next_to!(@users[3])
-      task.start!
-      task.review!
-      task.complete!(status: :routed_to_arc)
-    end
-
-    task = EstablishClaim.assign_next_to!(@users[4])
-
-    # assigning and moving the task to complete for
-    # user at index 5
-    3.times do |_index|
-      task = EstablishClaim.assign_next_to!(@users[5])
-      task.start!
-      task.review!
-      task.complete!(status: :routed_to_arc)
-    end
-
-    task = EstablishClaim.assign_next_to!(@users[6])
-
-    # Create one task with no decision documents
-    EstablishClaim.create(
-      appeal: tasks[2].appeal,
-      created_at: 5.days.ago
-    )
-
-    @tasks.push(*tasks)
-  end
-
-  def create_default_users
-    @users.push(
-      User.create(
-        css_id: "Reader",
-        station_id: "405",
-        full_name: "Padma VBMSStationIDMapsToMultipleVACOLSIDs Brannon"
-      )
-    )
-    @users.push(User.create(css_id: "Invalid Role", station_id: "283", full_name: "Cave InvalidRole Johnson"))
-    @users.push(User.create(css_id: "Establish Claim", station_id: "283", full_name: "Jane EstablishClaim Smith"))
-    @users.push(User.create(css_id: "Establish Claim 2", station_id: "405", full_name: "Carole EstablishClaim Johnson"))
-    @users.push(User.create(css_id: "Manage Claim Establishment", station_id: "283", full_name: "John ManageClaimEstablishment Doe"))
-    @users.push(User.create(css_id: "Certify Appeal", station_id: "283", full_name: "John CertifyAppeal Smith"))
-    @users.push(User.create(css_id: "System Admin", station_id: "283", full_name: "Angelina SystemAdmin Smith"))
-    @users.push(User.create(css_id: "Reader 2", station_id: "283", full_name: "Angelina ReaderAccess Smith"))
-    @users.push(User.create(css_id: "Hearing Prep", station_id: "283", full_name: "Lauren HearingPrep Roth"))
-    @users.push(User.create(css_id: "Mail Intake", station_id: "283", full_name: "Kwame MailIntake Nkrumah"))
-    @users.push(User.create(css_id: "Admin Intake", station_id: "283", full_name: "Ash AdminIntake Ketchum"))
-  end
 
   def create_annotations
     Generators::Annotation.create(comment: "Hello World!", document_id: 1, x: 300, y: 400)
     Generators::Annotation.create(comment: "This is an example comment", document_id: 2)
-  end
-
-  def create_ramp_elections(number)
-    number.times do |i|
-      RampElection.create!(
-        veteran_file_number: "#{i + 1}5555555",
-        notice_date: 1.week.ago
-      )
-    end
-
-    %w[11555555 12555555].each do |i|
-      RampElection.create!(
-        veteran_file_number: i,
-        notice_date: 3.weeks.ago
-      )
-    end
   end
 
   def create_tags
@@ -258,10 +33,6 @@ class SeedDB
       tag_id: Generators::Tag.create(text: "Right Knee").id,
       document_id: 2
     )
-  end
-
-  def create_api_key
-    ApiKey.new(consumer_name: "PUBLIC", key_string: "PUBLICDEMO123").save!
   end
 
   def create_higher_level_review_tasks
